requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
datefinder==0.7.3
DateTime==5.5
decorator==4.4.2
distro==1.9.0
easygui==0.98.3
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
instaloader==4.13.1
Jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
MarkupSafe==2.1.5
moviepy==1.0.3
mpmath==1.3.0
mutagen==1.47.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
openai==1.50.2
packaging==24.1
pillow==10.4.0
proglog==0.1.10
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pystyle==2.9
pyTelegramBotAPI==4.23.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.13.3
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
urllib3==2.2.2
uuid==1.30
websocket-client==1.8.0
websockets==12.0
yt-dlp==2024.5.27
zope.interface==7.0.3
